+++ packages/carlin/src/deploy/baseStack/command.ts
@@ -1,5 +1,4 @@
 import { CommandModule } from 'yargs';
-
 import { deployBaseStack } from './deployBaseStack';
 
 export const deployBaseStackCommand: CommandModule = {