+++ terezinha-farm/app/src/App.tsx
@@ -1,7 +1,10 @@
 // import * as React from 'react';
 import { Auth, useAuth } from '@ttoss/react-auth';
-import { Button, Stack } from '@ttoss/ui';
+import { Box, Button, Flex, Stack } from '@ttoss/ui';
 import { Markdown } from '@ttoss/components';
+
+import { FarmCorrectPagination } from './modules/Farm/FarmCorrectPagination';
+import { FarmWrongPagination } from './modules/Farm/FarmWrongPagination';
 
 const markdown = '# ~Hi~, *Pluto*!';
 
@@ -24,6 +27,16 @@ export const App = () => {
         Logout
       </Button>
       <Markdown>{markdown}</Markdown>
+
+      <Flex sx={{ paddingX: 'xl' }}>
+        <Box sx={{ flex: 1 }}>
+          <FarmWrongPagination />
+        </Box>
+
+        <Box sx={{ flex: 1 }}>
+          <FarmCorrectPagination />
+        </Box>
+      </Flex>
     </Stack>
   );
 };