+++ packages/forms/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.21.9](https://github.com/ttoss/ttoss/compare/@ttoss/forms@0.21.8...@ttoss/forms@0.21.9) (2023-12-12)
+
+### Bug Fixes
+
+- upload template when size is big ([#453](https://github.com/ttoss/ttoss/issues/453)) ([59371b7](https://github.com/ttoss/ttoss/commit/59371b7eb9befa006ffe316541f60dfc534dcbd4))
+
+## [0.21.8](https://github.com/ttoss/ttoss/compare/@ttoss/forms@0.21.7...@ttoss/forms@0.21.8) (2023-12-08)
+
+**Note:** Version bump only for package @ttoss/forms
+
 ## [0.21.7](https://github.com/ttoss/ttoss/compare/@ttoss/forms@0.21.6...@ttoss/forms@0.21.7) (2023-12-05)
 
 ### Bug Fixes