+++ packages/react-auth/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.20](https://github.com/ttoss/ttoss/compare/@ttoss/react-auth@1.7.19...@ttoss/react-auth@1.7.20) (2023-12-12)
+
+**Note:** Version bump only for package @ttoss/react-auth
+
+## [1.7.19](https://github.com/ttoss/ttoss/compare/@ttoss/react-auth@1.7.18...@ttoss/react-auth@1.7.19) (2023-12-08)
+
+**Note:** Version bump only for package @ttoss/react-auth
+
 ## [1.7.18](https://github.com/ttoss/ttoss/compare/@ttoss/react-auth@1.7.17...@ttoss/react-auth@1.7.18) (2023-12-05)
 
 **Note:** Version bump only for package @ttoss/react-auth